Transaction command_id type reduced from 8B to 2B
Summary: This reduces the size of a single `mvcc::Record<T>` from 40B to 24B.

Reviewers: buda, mislav.bradac, dgleich

Reviewed By: mislav.bradac, dgleich

Subscribers: mferencevic, pullbot

Differential Revision: https://phabricator.memgraph.io/D920

+++ src/transactions/type.hpp
@@ -8,5 +8,5 @@ namespace tx {
   using transaction_id_t = uint64_t;
 
   /** Type of a tx::Transcation's command id member */
-  using command_id_t = uint64_t;
+  using command_id_t = uint16_t;
 }
